initial create endpoint like

+++ src/business/post/PostBusiness.ts
@@ -4,33 +4,61 @@ import { Post } from "../../models/posts/Post";
 import {
   CreatePostInputDTO,
   CreatePostOutputDTO,
+  PostDB,
 } from "../../DTOs/posts_DTOs/create_Post_DTOS";
 import {
   edit_postDTOInput,
   edit_postDTOOutput,
 } from "../../DTOs/posts_DTOs/edit_post.DTO";
-import { PostController } from "../../controller/PostController";
+import { IdGenerator } from "../../services/IdGenerator";
+import { TokenManager } from "../../services/TokenManager";
+import { GetPostInputDTO } from "../../DTOs/posts_DTOs/get_posts_DTOs";
+import { BadRequestError } from "../../error/BadRequesteError";
 export class PostBusiness {
-  constructor(private postDatabase: PostDatabase) {}
-  public async get_post() {
+  constructor(
+    private postDatabase: PostDatabase,
+    private idGenerator: IdGenerator,
+    private tokenManager: TokenManager
+  ) {}
+  public async get_post(input: GetPostInputDTO) {
+    const {token} = input
+
+    const payload = this.tokenManager.getPayload(token)
+
+    if (payload === null) {
+      throw new BadRequestError("token inválido");
+    }
+
+
+
     const result = await this.postDatabase.get_post();
     return result;
   }
   public async create_post(
     input: CreatePostInputDTO
   ): Promise<CreatePostOutputDTO> {
-   
+    const { token } = input;
+    const isId = await this.postDatabase.get_user_by_id(input.creator_id);
+    if (isId) {
+      throw new NotFoundError("id já cadastrado");
+    }
+
+    const payload = this.tokenManager.getPayload(token)
+
+    if(payload === null){
+      throw new BadRequestError("token invalido")
+    }
     const new_post = new Post(
-      input.id,
+      this.idGenerator.generate(),
       input.creator_id,
       input.content,
-      input.likes,
-      input.dislikes,
-      input.created_at,
-      input.update_at
+      0,
+      0,
+      new Date().toISOString(),
+      new Date().toISOString()
     );
 
-    const new_post_db: CreatePostInputDTO = {
+    const new_post_db: PostDB = {
       id: new_post.get_id(),
       creator_id: new_post.get_creator_id(),
       content: new_post.get_content(),
@@ -52,15 +80,31 @@ export class PostBusiness {
   public async edit_post(
     input: edit_postDTOInput
   ): Promise<edit_postDTOOutput> {
+    const {token} = input
+    const payload = this.tokenManager.getPayload(token)
+    if(payload === null){
+      throw new BadRequestError("token invalido")
+    }
     const postDB = await this.postDatabase.get_post_by_id(input.id);
+
+    const postUpdate = new Post(
+      postDB.id,
+      postDB.creator_id,
+      input.content,
+      postDB.likes,
+      postDB.dislikes,
+      postDB.created_at,
+      new Date().toISOString()
+    );
+
     const post_update = {
-      id: postDB.id,
-      creator_id: postDB.creator_id,
-      content: input.content,
-      likes: postDB.likes,
-      dislikes: postDB.dislikes,
-      created_at: postDB.created_at,
-      update_at: postDB.update_at,
+      id: postUpdate.get_id(),
+      creator_id: postUpdate.get_creator_id(),
+      content: postUpdate.get_content(),
+      likes: postUpdate.get_likes(),
+      dislikes: postUpdate.get_dislikes(),
+      created_at: postUpdate.get_created_at(),
+      update_at: postUpdate.get_update_at(),
     };
     await this.postDatabase.edit_post(post_update);
     return {
@@ -70,8 +114,13 @@ export class PostBusiness {
 
   public async delete(id: string) {
     const post = await this.postDatabase.get_post_by_id(id);
+    console.log(post)
     if (!post) {
       throw new NotFoundError("post não encontrado");
+    }
+    const isUser = await this.postDatabase.checkIdUserInPost(post.creator_id)
+    if(isUser.length < 1){
+      throw new BadRequestError("usuario não criou este post");
     }
     const result = await this.postDatabase.delete_post(id);
     const output = {
@@ -79,5 +128,4 @@ export class PostBusiness {
     };
     return output;
   }
-  
 }
